+++ src/main/java/com/syqu/shop/service/ProductService.java
@@ -17,5 +17,5 @@ public interface ProductService {
     Page<Product> findByNameContaining(String name, Pageable pageable);
     // Page<Product> findAllSorted(String sortBy, Pageable pageable); // Add this method for sorting
     long count();
-    Page<Product> getFilteredProducts(int page, String sortBy, Integer categories, BigDecimal minPrice, BigDecimal maxPrice);
+    Page<Product> getFilteredProducts(int page, String sortBy, Long categories, BigDecimal minPrice, BigDecimal maxPrice);
 }
